Default more presets to double-decker layout
Breaking News, News Crawl, Sports, Business, Weather, Election and
Alert now default to the double (headline strip + crawl bar) layout
— it reads better in the listicle previews and matches the broadcast
reference more people expect. Minimal, Lower Third Crawl, and Custom
stay single-row since a second strip works against their purpose
(compact/minimal, or blank-slate).

+++ src/data/presets.ts
@@ -51,7 +51,7 @@ export const PRESETS: PresetDef[] = [
     label: 'News Crawl',
     description: 'Classic broadcast navy crawl bar',
     appearance: {
-      layout: 'single',
+      layout: 'double',
       background: 'solid',
       backgroundColor: '#111827',
       gradientFrom: '#1e3a8a',
@@ -72,7 +72,7 @@ export const PRESETS: PresetDef[] = [
     label: 'Sports',
     description: 'Energetic green/dark scoreboard style',
     appearance: {
-      layout: 'single',
+      layout: 'double',
       background: 'gradient',
       backgroundColor: '#052e16',
       gradientFrom: '#166534',
@@ -93,7 +93,7 @@ export const PRESETS: PresetDef[] = [
     label: 'Business',
     description: 'Muted grey, editorial, data-forward',
     appearance: {
-      layout: 'single',
+      layout: 'double',
       background: 'solid',
       backgroundColor: '#1f2937',
       gradientFrom: '#374151',
@@ -114,7 +114,7 @@ export const PRESETS: PresetDef[] = [
     label: 'Weather',
     description: 'Cool blue gradient, light text',
     appearance: {
-      layout: 'single',
+      layout: 'double',
       background: 'gradient',
       backgroundColor: '#0c4a6e',
       gradientFrom: '#0284c7',
@@ -135,7 +135,7 @@ export const PRESETS: PresetDef[] = [
     label: 'Election',
     description: 'Bold red/blue civic contrast',
     appearance: {
-      layout: 'single',
+      layout: 'double',
       background: 'solid',
       backgroundColor: '#1e1b4b',
       gradientFrom: '#312e81',
@@ -198,7 +198,7 @@ export const PRESETS: PresetDef[] = [
     label: 'Alert',
     description: 'High-contrast amber alert banner',
     appearance: {
-      layout: 'single',
+      layout: 'double',
       background: 'solid',
       backgroundColor: '#78350f',
       gradientFrom: '#f59e0b',
